refactor: mechanical warning cleanup, tail sweep to zero (Phase 5 T5)
Standard-treatment fixes (delete-or-use, drop meaningless top-level
const, default-or-delete, explicit initializers) for the remaining
AppleClang + g++14 warning tail, driving both compilers to census zero
alongside the real-bug/hardening commit. ~90 sites, ~90% dead scratch
locals (found/begin combo-inits, choice = 0, room_number = ch->in_room,
unused str[N] buffers, dead list "prev" trackers nothing reads) safe to
delete with zero behavior change.

Two sites required RNG-discipline care rather than a blind delete:
mage.cpp's caster_level = get_mage_caster_level(...) (x2) and mystic.cpp's
level = get_mystic_caster_level(...) both call functions with a number()
side effect (intel/will rounding) -- kept the calls, marked
[[maybe_unused]] instead of removing them. mage.cpp's
bool saved = new_saves_spell(...) call is similarly kept (its own
comment: "Just run through new_saves_spell for logging") for its
number() draw and spllog_* logging-global writes; only the unused
`saved` binding was dropped.

-Wignored-qualifiers (40 instances, single root cause): char_utils.h/.cpp's
`const bool is_fighting(...)` -- top-level const on a by-value bool
return has no effect; dropped at both declaration and definition.

-Wsign-compare (4): int-vs-size_type loop bounds in act_othe.cpp and
skill_timer.cpp (x3), static_cast<int>() on the container size.

-Wmissing-field-initializers (3): protocol.cpp's MSSP_t table entries
made their already-implicit nullptr trailing field explicit.

-Wunused-function (2): profs.cpp's operator<= (std::sort only needs
operator<) and spell_pa.cpp's get_casting_level (zero callers anywhere
in the tree) -- both deleted with proof of zero callers.

-Wunused-const-variable (2): modify.cpp's skill_fields[] and
act_info.cpp's sky_look[6] -- genuinely dead tables, deleted (act_info.cpp
lands in this commit for this one hunk; its other hunks are real-bug
fixes in the prior commit).

-Wunused-private-field (1): warrior_spec_handlers.h's
battle_mage_handler::character, never set or read.

-Wunused-label (1): act_offe.cpp's dead `delay:` label in do_kick, no
goto targets it anywhere in the file.

-Wtrigraphs (1): act_info.cpp's "(???)" literal contains the ??)
trigraph; escaped one '?' for identical resulting bytes.

-Wunknown-pragmas (66, g++14-only): GCC does not silently ignore
#pragma clang diagnostic ... the way earlier tasks' comments assumed --
it warns on every line. Made every push/ignored/pop trio in
act_info.cpp/script.cpp/utility.cpp portable via
#if defined(__clang__) ... #elif defined(__GNUC__) ... #endif (those
two files' pragma edits land in the prior real-bug commit alongside
their other hunks).

Full disposition table, census before/after both compilers, and gate
evidence: .superpowers/sdd/task-5-report.md (not committed, gitignored).

+++ src/skill_timer.cpp
@@ -28,7 +28,7 @@ void skill_timer::add_skill_timer(const char_data& ch, const int skill_id, const
 int skill_timer::report_skill_status(int player_id, char* buffer)
 {
     char str[255];
-    for (int i = 0; i < m_skill_timer.size(); ++i) {
+    for (int i = 0; i < static_cast<int>(m_skill_timer.size()); ++i) {
         auto& data = m_skill_timer[i];
 
         if (data.player_id == player_id && data.skill_id != GLOBAL_SKILL) {
@@ -41,7 +41,7 @@ int skill_timer::report_skill_status(int player_id, char* buffer)
 
 void skill_timer::update_skill_timer()
 {
-    for (int i = 0; i < m_skill_timer.size(); ++i) {
+    for (int i = 0; i < static_cast<int>(m_skill_timer.size()); ++i) {
         auto& data = m_skill_timer[i];
         if (data.counter > 0) {
             data.counter -= 1;
@@ -68,7 +68,7 @@ bool skill_timer::is_skill_allowed(const char_data& ch, const int skill_id)
 
     int player_id = utils::get_idnum(ch);
 
-    for (int i = 0; i < m_skill_timer.size(); ++i) {
+    for (int i = 0; i < static_cast<int>(m_skill_timer.size()); ++i) {
         auto data = m_skill_timer[i];
         if (data.player_id == player_id && (data.skill_id == skill_id || data.skill_id == GLOBAL_SKILL)) {
             return false;
